mejoras

+++ Juan_AmparoAP1_P1/Models/EntradasHuacales.cs
@@ -1,11 +1,11 @@
 using System.ComponentModel.DataAnnotations;
 using Juan_AmparoAP1_P1.DAL;
-namespace Juan_AmparoAP1_P1.Models 
+namespace Juan_AmparoAP1_P1.Models
 {
     public class EntradasHuacales
     {
         [Key]
-        public int idEntrada{ get; set; }
+        public int idEntrada { get; set; }
 
         [Required(ErrorMessage = "La fecha es obligatoria")]
         public DateTime Fecha { get; set; } = DateTime.Now;
@@ -14,12 +14,9 @@ namespace Juan_AmparoAP1_P1.Models
         public string? NombreCliente { get; set; }
 
         [Required(ErrorMessage = "La Cantidad es obligatoria")]
-        public double Cantidad { get; set; }
+        public int Cantidad { get; set; }
 
         [Required(ErrorMessage = "El Precio es Obligatorio")]
-        public double Precio { get; set; }
+        public int Precio { get; set; }
     }
 }
-
-
-
